refactor: validation rules
